requirements.txt
Django==5.2.6
pyTelegramBotAPI==4.29.1
pillow==11.3.0
python-dotenv==1.1.0
requests==2.32.5
gunicorn==23.0.0
whitenoise==6.9.0
dj-database-url==2.3.0
